+++ test/mds/cbht.c
@@ -3,7 +3,7 @@
  *                           <macan@ncic.ac.cn>
  *
  * Armed with EMACS.
- * Time-stamp: <2011-02-12 10:12:05 macan>
+ * Time-stamp: <2011-05-10 12:36:31 macan>
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -52,7 +52,7 @@ void hmr_print(struct hvfs_md_reply *hmr)
     hvfs_info(mds, "hmr-> HI: len %d, flag 0x%x, uuid %lx, hash %ld, "
               "itbid %ld, puuid %ld, psalt %ld\n", 
               hi->namelen, hi->flag, hi->uuid, hi->hash, 
-              hi->itbid, hi->puuid, hi->psalt);
+              (u64)hi->itbid, hi->puuid, hi->psalt);
     p += sizeof(struct hvfs_index);
     if (hmr->flag & MD_REPLY_WITH_MDU) {
         m = (struct mdu *)p;
